+++ examples/04_saxpy/serial.cpp
@@ -46,9 +46,9 @@ int main(int argc, char* argv[])
 
     options_description desc_commandline;
     desc_commandline.add_options()
-        ("N", value<std::uint64_t>()->default_value(1024),
+        ("N", value<std::size_t>()->default_value(1024),
          "Number of Elements")
-        ("steps", value<std::uint64_t>()->default_value(100),
+        ("steps", value<std::size_t>()->default_value(100),
          "Number of steps to apply the stencil")
     ;
